requirements.txt
pytest-playwright==0.7.0
pytest-html==4.1.1